docs/curso-a2a-helloworld/helloworld/pyproject.toml
[project]
name = "helloworld"
version = "0.2.0"
description = "HelloWorld agent completo - migração de hello_world + funcionalidades avançadas"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[sqlite]>=0.2.6",
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    # Dependências adicionais para funcionalidades avançadas
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true